refactor setup functions to replay_setup
